requirements.txt
pyarrow==18.1.0
pandas
faker
unitycatalog-client
python-dotenv